+++ README.md
@@ -1,5 +1,5 @@
 # spring-multi-module-template
 SpringBoot 多模块项目模版。使用buildSrc来管理公共依赖。默认启用graalvm插件。
 
-默认配置： SpringBoot 3.2.1 + Kotlin 1.9.20 + graalvm-native-plugin + JDK17
+默认配置： SpringBoot 3.3.0 + Kotlin 2.0.0 + JDK17
