--- api/src/main/kotlin/io/github/llh4github/smmtemplate/api/ApiApp.kt
@@ -1,17 +1,0 @@
-package io.github.llh4github.smmtemplate.api
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-/**
- *
- * Created At 2023/12/29 22:49
- * @author llh
- */
-@SpringBootApplication
-class ApiApp
-
-
-fun main(args: Array<String>) {
-    runApplication<ApiApp>(*args)
-}
