--- api/src/main/kotlin/io/github/llh4github/smmtemplate/api/api/IndexApi.kt
@@ -1,28 +1,0 @@
-package io.github.llh4github.smmtemplate.api.api
-
-import io.github.llh4github.smmtmplate.commons.JsonWrapper
-import io.github.oshai.kotlinlogging.KotlinLogging
-import io.swagger.v3.oas.annotations.Operation
-import io.swagger.v3.oas.annotations.tags.Tag
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RestController
-import java.time.LocalDateTime
-
-/**
- *
- * Created At 2023/12/30 21:28
- * @author llh
- */
-@Tag(name = "测试接口")
-@RestController
-class IndexApi : BaseApi() {
-
-    private val logger = KotlinLogging.logger {}
-
-    @Operation(summary = "获取当前时间")
-    @GetMapping("now")
-    fun now(): JsonWrapper<LocalDateTime> {
-        logger.debug { "调用测试" }
-        return ok(LocalDateTime.now())
-    }
-}
