--- api/src/test/kotlin/io/github/llh4github/smmtemplate/api/ApiAppTest.kt
@@ -1,17 +1,0 @@
-package io.github.llh4github.smmtemplate.api
-
-import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-/**
- *
- * Created At 2023/12/30 10:46
- * @author llh
- */
-@SpringBootTest
-class ApiAppTest{
-
-    @Test
-    fun context(){}
-}
