+++ buildSrc/src/main/kotlin/spring-conventions.gradle.kts
@@ -5,9 +5,4 @@ plugins {
 apply(plugin = "org.jetbrains.kotlin.plugin.spring")
 apply(plugin = "org.springframework.boot")
 apply(plugin = "io.spring.dependency-management")
-apply(plugin = "org.graalvm.buildtools.native")
 
-springBoot {
-    // Creates META-INF/build-info.properties for Spring Boot Actuator
-    buildInfo()
-}
