--- commons/build.gradle.kts
@@ -1,9 +1,0 @@
-
-plugins {
-    id("kotlin-conventions")
-    `java-library`
-}
-
-dependencies {
-    api(libs.swagger.annotations)
-}
