+++ gradle.properties
@@ -6,7 +6,6 @@ version=0.1.0-SNAPSHOT
 org.gradle.daemon=true
 org.gradle.caching=true
 org.gradle.parallel=true
-#????
 org.gradle.configureondemand=true
