+++ mod-commons/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    id("kotlin-conventions")
+    `java-library`
+}
+
+dependencies {
+    compileOnly(libs.swagger.annotations)
+    compileOnly(libs.slf4j.api)
+    compileOnly(libs.logback.core)
+    compileOnly(libs.logback.classic)
+}
