+++ mod-commons/src/main/kotlin/Main.kt
@@ -1,0 +1,5 @@
+package io.github.llh4github.smmtemplate
+
+/**
+ * 占位用，除web模块外，都不建议在根目录下直接创建类
+ */
