+++ mod-web/build.gradle.kts
@@ -1,0 +1,16 @@
+plugins {
+    id("kotlin-conventions")
+    id("spring-conventions")
+    alias(libs.plugins.ksp)
+}
+
+dependencies {
+    implementation("org.springframework.boot:spring-boot-starter-web")
+    implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
+    implementation("org.jetbrains.kotlin:kotlin-reflect")
+    annotationProcessor("org.springframework.boot:spring-boot-configuration-processor")
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+
+    implementation(libs.knife4j.openapi3)
+    implementation(project(":mod-commons"))
+}
