+++ mod-web/src/main/kotlin/Main.kt
@@ -1,0 +1,11 @@
+package io.github.llh4github.smmtemplate
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class WebMain
+
+fun main(args: Array<String>) {
+    runApplication<WebMain>(*args)
+}
