+++ mod-web/src/test/kotlin/MainTest.kt
@@ -1,0 +1,12 @@
+package io.github.llh4github.smmtemplate
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class MainTest {
+
+    @Test
+    fun contextLoads() {
+    }
+}
