+++ settings.gradle.kts
@@ -1,8 +1,8 @@
-
 plugins {
     // Apply the foojay-resolver plugin to allow automatic download of JDKs
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.7.0"
 }
 
 rootProject.name = "spring-multi-module-template"
-include("api", "commons")
+include("mod-commons")
+include("mod-web")
